+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_user_service_functions.go
@@ -343,6 +343,29 @@ func (backend *DockerKurtosisBackend) GetConnectionWithUserService(
 	return newConnection, nil
 }
 
+func (backend *DockerKurtosisBackend) CopyFromUserService(
+	ctx context.Context,
+	enclaveId enclave.EnclaveID,
+	serviceGuid service.ServiceGUID,
+	filepathInService string,
+)(
+	io.ReadCloser,
+	error,
+) {
+
+	userServiceContainerId, _, err := backend.getSingleUserService(ctx, enclaveId, serviceGuid)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting user service with GUID '%v' in enclave with ID '%v'", serviceGuid, enclaveId)
+	}
+
+	tarStreamReadCloser, err := backend.dockerManager.CopyFromContainer(ctx, userServiceContainerId, filepathInService)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred copying file with filepath '%v' from user service with GUID '%v' and container ID '%v' ", filepathInService, serviceGuid, userServiceContainerId)
+	}
+
+	return tarStreamReadCloser, nil
+}
+
 func (backend *DockerKurtosisBackend) StopUserServices(
 	ctx context.Context,
 	filters *service.ServiceFilters,